Move expr tests
